+++ src/app/models/localstorage/Settings/gui.js
@@ -29,6 +29,7 @@ export const ATTR_GUI_FOCUSREFRESH_FIVE = 300000;
 export default Fragment.extend({
 	externalcommands: attr( "boolean", { defaultValue: false } ),
 	focusrefresh: attr( "number", { defaultValue: ATTR_GUI_FOCUSREFRESH_NONE } ),
+	hidebuttons: attr( "boolean", { defaultValue: false } ),
 	homepage: attr( "string", { defaultValue: "/featured" } ),
 	integration: attr( "number", { defaultValue: ATTR_GUI_INTEGRATION_BOTH } ),
 	minimize: attr( "number", { defaultValue: ATTR_GUI_MINIMIZE_NOOP } ),